Separated myBindHelper from main
